Improved check for plugin routing
An app plugin with 'frontend' config option enabled does not necesarily have to contain 'routing.php' config file and can instead update app routing by means of an event handler in its main class. This fix assumes that such an event is always called 'routing' in all apps (which does not have to be true but will most probably be so in 99% cases) and improves routing handling check for such plugins.

Now this CLI method says 'empty routing for frontend' when I try to compress my plugin, which works absolutely fine without a 'routing.php' file.